+++ lib/orchid/symbiont/application.ex
@@ -1,4 +1,4 @@
-defmodule Orchid.Symbiont.Application do
+defmodule OrchidSymbiont.Application do
   @moduledoc false
 
   use Application
@@ -6,15 +6,15 @@ defmodule Orchid.Symbiont.Application do
   @impl true
   def start(_type, _args) do
     children = [
-      {Registry, keys: :unique, name: Orchid.Symbiont.Registry},
-      Orchid.Symbiont.Catalog,
-      {Orchid.Symbiont.Runtime, session_id: nil},
+      {Registry, keys: :unique, name: OrchidSymbiont.Registry},
+      OrchidSymbiont.Catalog,
+      {OrchidSymbiont.Runtime, session_id: nil},
       # TODO: Add a `strict_mode`
       # If ture, not fallback to global mode.
-      {Task.Supervisor, name: Orchid.Symbiont.Preloader},
+      {Task.Supervisor, name: OrchidSymbiont.Preloader},
     ]
 
-    opts = [strategy: :one_for_one, name: Orchid.Symbiont.Supervisor]
+    opts = [strategy: :one_for_one, name: OrchidSymbiont.Supervisor]
     Supervisor.start_link(children, opts)
   end
 end