+++ lib/orchid/symbiont/catalog.ex
@@ -1,24 +1,24 @@
-defmodule Orchid.Symbiont.Catalog do
+defmodule OrchidSymbiont.Catalog do
   use GenServer
 
   @type sesssion_id:: binary()
-  @type name :: {sesssion_id(), Orchid.Symbiont.Step.symbiont_name()}
+  @type name :: {sesssion_id(), OrchidSymbiont.Step.symbiont_name()}
 
   def start_link(_opts) do
     GenServer.start_link(__MODULE__, %{}, name: __MODULE__)
   end
 
-  @spec register(name() | Orchid.Symbiont.Step.symbiont_name(), {module(), term()}) :: any()
+  @spec register(name() | OrchidSymbiont.Step.symbiont_name(), {module(), term()}) :: any()
   def register(name, {mod, args}) do
     GenServer.call(__MODULE__, {:register, name, {mod, args}})
   end
 
-  @spec lookup(Orchid.Symbiont.Step.symbiont_name()) :: any()
+  @spec lookup(OrchidSymbiont.Step.symbiont_name()) :: any()
   def lookup(name) do
     GenServer.call(__MODULE__, {:lookup, name})
   end
 
-  @spec lookup(sesssion_id(), Orchid.Symbiont.Step.symbiont_name()) :: any()
+  @spec lookup(sesssion_id(), OrchidSymbiont.Step.symbiont_name()) :: any()
   def lookup(session_id, name) when not is_nil(session_id) do
     GenServer.call(__MODULE__, {:lookup, {session_id, name}})
     |> case do